Update from htmlunit/master

+++ src/test/java/com/gargoylesoftware/htmlunit/WebServerTestCase.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -264,7 +264,7 @@ public abstract class WebServerTestCase extends WebTestCase {
      * @return the page
      * @throws Exception if something goes wrong
      */
-    protected final HtmlPage loadPageWithAlerts(final String html, final URL url, final int maxWaitTime)
+    protected final HtmlPage loadPageWithAlerts(final String html, final URL url, final long maxWaitTime)
         throws Exception {
         alertHandler_.clear();
         expandExpectedAlertsVariables(URL_FIRST);
